last update

+++ ft_atoi.c
@@ -6,20 +6,19 @@
 /*   By: zlaarous <zlaarous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/16 14:25:21 by zlaarous          #+#    #+#             */
-/*   Updated: 2022/11/01 14:14:59 by zlaarous         ###   ########.fr       */
+/*   Updated: 2022/11/01 20:38:53 by zlaarous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static void	space(const char *str, int *i)
+static void space(const char *str, int *i)
 {
-	while (str[*i] == ' ' || str[*i] == '\t' || str[*i] == '\v'
-		|| str[*i] == '\r' || str[*i] == '\f' || str[*i] == '\n')
+	while (str[*i] == ' ' || str[*i] == '\t' || str[*i] == '\v' || str[*i] == '\r' || str[*i] == '\f' || str[*i] == '\n')
 		(*i)++;
 }
 
-static void	sign_in(char const *str, int *i, int *sign)
+static void sign_in(char const *str, int *i, int *sign)
 {
 	if (str[*i] == '-')
 	{
@@ -30,12 +29,12 @@ static void	sign_in(char const *str, int *i, int *sign)
 		(*i)++;
 }
 
-int	ft_atoi(const char *str)
+int ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
-	int	previous_result;
+	int i;
+	int sign;
+	long long result;
+	long long previous_result;
 
 	sign = 1;
 	result = 0;
@@ -58,4 +57,3 @@ int	ft_atoi(const char *str)
 	}
 	return (sign * result);
 }
-
